Add files via upload
This is my first lab in FPGA class.
For this lab each student was given a unique 4-input boolean function.
My function was y = f ( x_3, x_2, x_1, x_0 ) = V ( 1, 4, 5, 6, 7, 8, 10, 12, 14 ) or otherwise 0101010111110010.

Task 1: minimize the given function using Karnaugh Maps and then describe the result in Verilog using transistors.

Task 2: describe the given function in Verilog as a PROM curcuit using transistors.

Task 3. describe the given function in Verilog using built-in primitives.
a) based on XOR-AND logic;
b) based on NOR logic.
c) describe the given function in Verilog with a user defined primitive.

Also we had to make an additional Enable input for each task.

The results of simulations as well as the drawings for all the circuits can be found in the Lab Report file that is also added in this commit (pages 12 - 24).