Add files via upload
This is a lab report file. It has detailed info about both the homework that had to be done prior to the lab and the lab itself.
There is all of the math like minimization of my boolean functions and also there are screenshots of the simulation results in ModelSim and the photos of my development board that I was running the code at.