Add files via upload
This is my sixth lab in the FPGA class.

In this lab we had to describe in SystemVerilog a sequence detector.
Every student was given a unique sequence to work with. My sequence ended up being 1100.

We were also asked to make this sequence detector with both Moore FSM and Mealy FSM.
Both Moore and Melay FSM we had to describe in six different ways:
- based on D-, T-, RS-, JK- flip-flops
  where you can choose a flip-flop type with a parameter
- states in binary code,
- states in unary code,
- states in Gray code,
- states in Johnson code,
- based on a shift register.

Additionally in order to generate the pseudorandom sequence inside which a given subsequence could be detected, I made a Galois.sv module. In that module I described 16-bit a pseudorandom sequence generator based on a Galois LFSR.

This project was first simulated in ModelSim and then run on my Altera Cyclone IV development board.

There's also a lab report file in this branch where more detailed info can be found.