Add files via upload
This is my crossroads project.

In this project I've created a physical model of crossroads with LED traffic lights and with a boom gate that I made with a servo actuator. I've programmed the traffic lights using a Finite State Machine. The servo for a boom gate is controlled with PWM signal so I also had to include a PWM generator in my code. The core of this project was my development board loaded with Altera Cyclone IV FPGA.

In this branch I've also included my project report where more detailed info and photos of my model can be found.